add a panicing test case to quadtree/quadtree_test.go/TestQuadTreeMatching
this test case panics because in quadtree/quadtree.go does not check
if the `findVisitor.closest` value is nil before returning
`v.closest.Value` at the end of the method.

a fix here is not straight forward because the method signature does not
return `error`. It's possible that we could add a `type orb.NilPointer`
[ ugh naming ] that satisfies the orb.Poitner interface but could be
returned without breaking existing clients