hacky fix to Matching(…) panic
just return a private, nilPointer struct that implements orb.Pointer by
returning an orb.Point{} directly. This change is technically breaking
in that it no longer reliably panics if users were relying on that
behavior. It is the simplest change I could think of that didn't change
the API.

Alternatively, we can introduce new FindV2 and MatchingV2 methods that
return (orb.Pointer, error) that do not panic that can survive until a
minor orb version bump with breaking API.